safari-534.53-branch/LayoutTests/http/tests/security/dataURL/xss-DENIED-to-data-url-window-open-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL data:text/html,%3Chtml%3E%3Chead%3E%3Cscript%3Efunction%20fireSentinel()%20{if%20(window.layoutTestController)layoutTestController.globalFlag%20=%20true;}%3C/script%3E%3C/head%3E%3Cbody%20onload=%22fireSentinel();%22%3E%3Cp%3EOpened%20Frame%3C/p%3E%3Cp%20id='accessMe'%3EPASS:%20Cross%20frame%20access%20from%20an%20opener%20frame%20was%20denied%3C/p%3E%3C/body%3E%3C/html%3E from frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-to-data-url-window-open.html. Domains, protocols and ports must match.

Opener Frame

Test
